requirements.txt
streamlit
yfinance
pandas
requests
pyotp
smartapi-python